WorldDataHandler v2
Some minor improvements to the structure of the data.

We use an array of BasePixel instead of an array of pointers, this is in an attempt to make the pixel data as close together as possible. Previously they were constructed using new and stored as pointers so while there was a good chance they were together in memory there was no guarantee.

Now as a collection where they have to be together, we should avoid cache misses more and get a fair bit of an improvement. The class and the Pixels are all constexpr constructors.