Many big changes (lazy commit due to period of time), debug build can now run again
Months of previous changes I've long forgotten, and a few updates to compiler definitions such as _DEBUG (which is used by imgui) to PIXEL_PROJECT_DEBUG which strips out debug messages.

Removal of fmt which was causing problems, using modern c++ std::format which does most of the same, plus the tweaks to make this change work with project.